parse more header files and skips ahead to synchronize if the data is corrupted

+++ launch_file/src/lib.rs
@@ -1,10 +1,11 @@
 mod deserialize;
 mod bytes;
 
-use std::sync::Arc;
+use std::sync::{Arc, LazyLock};
 use std::ffi::OsStr;
 use std::{fs, fs::File};
 use std::{io, io::{Read, Write}};
+use std::io::Seek;
 use std::num::NonZeroU32;
 use std::path::Path;
 use std::process::Command;
@@ -31,6 +32,10 @@ impl Checksum {
     pub const SENTINEL: u32 = u32::from_le_bytes([0xDE, 0xAD, 0xBE, 0xEF]);
 }
 
+static SCRIPT_DIR: LazyLock<Option<ProjectDirs>> = LazyLock::new(|| {
+    ProjectDirs::from("", "", "MIDAS Launch")
+});
+
 
 #[derive(Deserialize, Clone, Eq, PartialEq)]
 pub struct LogFormat {
@@ -42,10 +47,11 @@ pub struct LogFormat {
 
 impl LogFormat {
     pub fn clear_scripts() {
-        let script_dir = ProjectDirs::from("", "", "MIDAS Launch").unwrap();
-        fs::create_dir_all(script_dir.data_dir()).unwrap();
-        let _ = fs::remove_file(script_dir.data_dir().join("__main__.py"));
-        let _ = fs::remove_file(script_dir.data_dir().join("cpp_parser.py"));
+        if let Some(script_dir) = SCRIPT_DIR.as_ref() {
+            fs::create_dir_all(script_dir.data_dir()).unwrap();
+            let _ = fs::remove_file(script_dir.data_dir().join("__main__.py"));
+            let _ = fs::remove_file(script_dir.data_dir().join("cpp_parser.py"));
+        }
     }
 
     pub fn from_inline_header(data: &[u8]) -> Result<Self, String> {
@@ -57,8 +63,7 @@ impl LogFormat {
     }
 
     pub fn from_format_file(format_file_name: &Path, python: impl AsRef<OsStr>) -> Result<Self, String> {
-        let script_dir = ProjectDirs::from("", "", "MIDAS-Launch")
-            .ok_or("Could not find script.".to_string())?;
+        let script_dir = SCRIPT_DIR.as_ref().ok_or("Could not find script.".to_string())?;
 
         fs::create_dir_all(script_dir.data_dir()).map_err(|e| format!("Could not create script: {}", e))?;
         fs::create_dir_all(script_dir.cache_dir()).map_err(|e| format!("Could not create script: {}", e))?;
@@ -141,12 +146,13 @@ impl<'f> LaunchFileReader<'f> {
         let mut smallest = usize::MAX;
         let mut largest = usize::MIN;
         for (name, (disc, format)) in &format.variants {
-            let mut builder = DeserializerBuilder::new(&mut dataframe_builder);
+            let mut builder = DeserializerBuilder::new(name.clone(), &mut dataframe_builder);
             format.to_fast(&mut builder, name);
             let fast_format = builder.finish();
             smallest = smallest.min(fast_format.size).max(1);
             largest = largest.max(fast_format.size);
 
+            dbg!(name, fast_format.size);
             let key = dataframe_builder.add_interned_string(name);
             variants.insert(*disc, (key, fast_format));
         }
@@ -171,7 +177,7 @@ impl<'f> LaunchFileReader<'f> {
         }
     }
 
-    pub fn read_file(&mut self, file: &mut impl Read, mut on_row_callback: impl FnMut(u64)) -> io::Result<u64> {
+    pub fn read_file(&mut self, file: &mut (impl Read + Seek), mut on_row_callback: impl FnMut(u64)) -> io::Result<u64> {
         // todo
         // if let Some(file_size) = file_size {
         //     let maximum_needed_rows = file_size.div_ceil(self.smallest as u64 + 8);
@@ -186,15 +192,35 @@ impl<'f> LaunchFileReader<'f> {
 
         let result: io::Result<()> = try_catch!({
             let mut read_buf = vec![0u8; self.largest].into_boxed_slice();
+            let mut last: Option<&Deserializer> = None;
+            let mut last_timestamp = 0;
+            let mut synchronizing_amount = 0;
             loop {
-                let row_idx = self.dataframe.add_null_row();
-                let mut row = self.dataframe.row_mut(row_idx);
-
                 let determinant = file.read_u32::<LittleEndian>()?; offset += 4;
                 let timestamp_ms = file.read_u32::<LittleEndian>()?; offset += 4;
 
-                let (key, fast_format) = self.variants.get(&determinant)
-                    .ok_or_else(|| io::Error::other(format!("No variant for discriminant {} at offset {}", determinant, offset - 8)))?;
+                let Some((key, fast_format)) = self.variants.get(&determinant) else {
+                    file.seek_relative(-7)?;
+                    offset -= 7;
+                    synchronizing_amount += 1;
+                    // return Err(io::Error::other(format!("No variant for discriminant {} at offset {}", determinant, offset - 4 + rewind_amount as u64)));
+                    continue;
+                };
+                if last_timestamp != 0 && timestamp_ms.abs_diff(last_timestamp) >= 500 {
+                    file.seek_relative(-7)?;
+                    offset -= 7;
+                    synchronizing_amount += 1;
+                    continue;
+                }
+                if synchronizing_amount != 0 {
+                    eprintln!("Stepped {} bytes forward from offset {} to synchronize to timestamp {}.", synchronizing_amount, offset -7 - synchronizing_amount, timestamp_ms);
+                    synchronizing_amount = 0;
+                }
+                last = Some(fast_format);
+                last_timestamp = timestamp_ms;
+
+                let row_idx = self.dataframe.add_null_row();
+                let mut row = self.dataframe.row_mut(row_idx);
 
                 row.set_col_raw(0, Some(*key));
                 row.set_col_with_ty(1, DataType::Integer, Data::Integer(self.file_number - 1));
